+++ components/Input.tsx
@@ -3,7 +3,7 @@ import { iInput } from "../src/types/props.ts";
 import setup from "../src/setup/Input.ts";
 
 /**
- * @todo Add custom classes support
+ * @todo Choose a standard color for placeholders.
  */
 export default function (props: Partial<iInput>) {
   const { c, maxWidth, label, error, ...p } = setup(props);