+++ components/TextArea.tsx
@@ -3,7 +3,7 @@ import { iTextArea } from "../src/types/props.ts";
 import setup from "../src/setup/TextArea.ts";
 
 /**
- * @todo Replace `<span>` with `<Text>`.
+ * @todo Choose a standard color for placeholders.
  */
 export default function TextArea(props: Partial<iTextArea>) {
   const { c, maxWidth, label, error, ...p } = setup(props);